+++ CHANGELOG.md
@@ -21,7 +21,8 @@ recommendations of [keepachangelog.com](http://keepachangelog.com/).
 
 ### Fixed
 
-- None
+- [#1238](https://github.com/paper-trail-gem/paper_trail/pull/1238) -
+  Query optimization in `reify`
 
 ### Dependencies